// frontend/src/app/core/api-client/model/updateRequest.ts
/**
 * Access management PoC
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RoleEnum, StatusEnum } from "./userRequest";


/**
 * Request DTO for updating a user
 */
export interface UpdateRequest { 
    /**
     * Email address of the user
     */
    email?: string;
    /**
     * Role assigned to the user
     */
    role?: RoleEnum;
    /**
     * Status of the user account
     */
    status?: StatusEnum;
}